% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accountbalance.R
\docType{data}
\name{accountbalance}
\alias{accountbalance}
\title{accountbalance}
\format{A data frame with 100 observations on the following 7 variables.

\describe{
\item{\code{Accbl}}{total amount of money in the account in the account balance}
\item{\code{Age}}{a numeric vector}
\item{\code{Fams}}{family size}
\item{\code{Income}}{a numeric vector}
\item{\code{Expenditure}}{a numeric vector}
\item{\code{Ocards}}{ownership of other cards}
\item{\code{Online}}{online purchases}
}}
\usage{
accountbalance
}
\description{
This dataset shows the account balance of 100 people with their own Age, Income, Family size, Expenditure,ownership of other cards and online purchases.
}
\keyword{datasets}
