% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import data}
\usage{
import_data()
}
\value{
dataframe
}
\description{
This function allows to import the dataset in the package.
Run this first function to recall the dataset
}
\examples{
import_data()
}
