% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.r
\name{linear.r}
\alias{linear.r}
\title{linear.r}
\usage{
linear.r()
}
\value{
The coefficients of the regression
}
\description{
this function allows to make a linear regression between Age and the explanatory variables(Age,Fams,Income,Expenditure)
}
\examples{
linear.r()




}
