% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearD.R
\name{linearD}
\alias{linearD}
\title{linearD}
\usage{
linearD()
}
\value{
the coefficient of the regression with the dummy variables
}
\description{
this function allows to allows to make a linear regression between Age and the explanatory variables(Age,Fams,Income,Expenditure with dummy Ocards and Online)
}
\examples{
 linearD()
}
