% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotres.R
\name{plotres}
\alias{plotres}
\title{This function plots the residual against the predicted model}
\usage{
plotres(x, y)
}
\description{
@param x is/are the observed variable/s that we predict using predict
 @param y are the residuals
 @import graphics
 @import stats
 @export
 @examples
 plotres(predict(linear.r), residuals(linear.r))
 plotres(predict(linearD), residuals(linearD))
}
