% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_output.R
\name{regression_output}
\alias{regression_output}
\title{regression_output}
\usage{
regression_output(x)
}
\arguments{
\item{x}{is the regression (linear.r or linearD)}
}
\description{
This function shows a more detailed output for each regression
}
\examples{
regression_output(linear.r())
regression_output(linearD())


}
