% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship.R
\name{relationship}
\alias{relationship}
\title{This function shows the relatioship between some variables and accountbalance}
\usage{
relationship(x)
}
\arguments{
\item{x}{is a name of a variable (Age, Income, Expenditure)}
}
\value{
a plot
}
\description{
This function shows the relatioship between some variables and accountbalance
}
\examples{
relationship(~Age)
relationship(~Income)
}
