% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{res}
\alias{res}
\title{res}
\usage{
res(x)
}
\arguments{
\item{x}{is the value of the function of the linear regression}
}
\value{
the residuals in ascending order
}
\description{
this functions allows you to see the residuals of the linear regression in ascending order
}
\examples{
res(linear.r())
res(linearD())
}
