% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotres.R
\name{plotres}
\alias{plotres}
\title{plotres}
\usage{
plotres(x)
}
\arguments{
\item{x}{is the dataset on which the function calculates the residual and the predicted value}
}
\description{
This function plots the residuals against the predicted model
}
\examples{






}
