% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{res}
\alias{res}
\title{res
 @description this functions allows you to see the residuals of the linear regression in ascending order
 @param x is the linear regression
 @export
 @examples
 res(linear.r())
 res(linearD())}
\usage{
res(x)
}
\description{
res
 @description this functions allows you to see the residuals of the linear regression in ascending order
 @param x is the linear regression
 @export
 @examples
 res(linear.r())
 res(linearD())
}
